siste finjusteringer
